requirements.txt
torchsummary
numpy
Pillow
opencv-python
tkinter
yaml
scipy
tqdm
lpips